update jump search

+++ 0x1E-search_algorithms/search_algos.h
@@ -12,6 +12,6 @@ int jump_search(int *array, size_t size, int value);
 
 void print_array(int *array, size_t start, size_t end);
 int binary_search_func(int *array, size_t start, size_t end, int value);
-size_t search_func(int *array, size_t i, size_t size, int value);
+int search_func(int *array, size_t i, size_t size, int value);
 
 #endif
